requirements.txt
instabot==0.4.5
requests==2.22.0
python-dotenv==0.10.3